throw in some 'just works' basic auth for now

+++ app/Http/Controllers/Web/WebhookController.php
@@ -7,12 +7,17 @@ use App\Models\WebhookSubscriber;
 use App\Models\WebhookSubscriberHeader;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\View\View;
 
 class WebhookController extends Controller
 {
     public function index(): View
     {
+        if (!Auth::user()->can_view_webhooks && !Auth::user()->can_manage_webhooks) {
+            abort(403);
+        }
+
         $webhooks = WebhookSubscriber::query()->get();
 
         return view('webhooks.index', compact('webhooks'));
@@ -20,11 +25,19 @@ class WebhookController extends Controller
 
     public function create(): View
     {
+        if (!Auth::user()->can_manage_webhooks) {
+            abort(403);
+        }
+
         return view('webhooks.create');
     }
 
     public function store(Request $request): RedirectResponse
     {
+        if (!Auth::user()->can_manage_webhooks) {
+            abort(403);
+        }
+
         $url = $request->get('url');
         $headers = array_combine(
             $request->get('headers', []),
@@ -46,11 +59,19 @@ class WebhookController extends Controller
 
     public function show(WebhookSubscriber $webhook): View
     {
+        if (!Auth::user()->can_manage_webhooks) {
+            abort(403);
+        }
+
         return view('webhooks.show', compact('webhook'));
     }
 
     public function destroy(WebhookSubscriber $webhook): RedirectResponse
     {
+        if (!Auth::user()->can_manage_webhooks) {
+            abort(403);
+        }
+
         $webhook->delete();
 
         return redirect(route('webhooks.index'));
